+++ packages/@expo/cli/src/start/server/metro/createServerComponentsMiddleware.ts
@@ -46,12 +46,17 @@ export function createServerComponentsMiddleware(
     ssrLoadModule,
     ssrLoadModuleArtifacts,
     useClientRouter,
+    createModuleId,
   }: {
     rscPath: string;
     instanceMetroOptions: Partial<ExpoMetroOptions>;
     ssrLoadModule: SSRLoadModuleFunc;
     ssrLoadModuleArtifacts: SSRLoadModuleArtifactsFunc;
     useClientRouter: boolean;
+    createModuleId: (
+      filePath: string,
+      context: { platform: string; environment: string }
+    ) => string | number;
   }
 ) {
   const routerModule = useClientRouter
@@ -116,7 +121,11 @@ export function createServerComponentsMiddleware(
   }
 
   async function exportServerActionsAsync(
-    { platform, entryPoints }: { platform: string; entryPoints: string[] },
+    {
+      platform,
+      entryPoints,
+      domRoot,
+    }: { platform: string; entryPoints: string[]; domRoot?: string },
     files: ExportAssetMap
   ): Promise<{
     clientBoundaries: string[];
@@ -124,10 +133,9 @@ export function createServerComponentsMiddleware(
   }> {
     const uniqueEntryPoints = [...new Set(entryPoints)];
     // TODO: Support multiple entry points in a single split server bundle...
-    const serverRoot = getMetroServerRootMemo(projectRoot);
-
     const manifest: Record<string, [string, string]> = {};
     const nestedClientBoundaries: string[] = [];
+
     for (const entryPoint of uniqueEntryPoints) {
       const contents = await ssrLoadModuleArtifacts(entryPoint, {
         environment: 'react-server',
@@ -136,6 +144,8 @@ export function createServerComponentsMiddleware(
         modulesOnly: true,
         // Required
         runModule: true,
+        // Required to ensure assets load as client boundaries.
+        domRoot,
       });
 
       const reactClientReferences = contents.artifacts
@@ -153,7 +163,11 @@ export function createServerComponentsMiddleware(
             entryPoint
         );
       }
-      const relativeName = path.relative(serverRoot, entryPoint);
+
+      const relativeName = createModuleId(entryPoint, {
+        platform,
+        environment: 'react-server',
+      });
       const safeName = path.basename(contents.artifacts.find((a) => a.type === 'js')!.filename!);
 
       const outputName = `_expo/rsc/${platform}/${safeName}`;
@@ -164,7 +178,7 @@ export function createServerComponentsMiddleware(
       });
 
       // Import relative to `dist/server/_expo/rsc/web/router.js`
-      manifest[entryPoint] = [relativeName, outputName];
+      manifest[entryPoint] = [String(relativeName), outputName];
     }
 
     // Save the SSR manifest so we can perform more replacements in the server renderer and with server actions.
@@ -243,7 +257,13 @@ export function createServerComponentsMiddleware(
     platform: string;
     engine?: 'hermes' | null;
     ssrManifest?: Map<string, string>;
-  }) {
+  }): (
+    file: string,
+    isServer: boolean
+  ) => {
+    id: string;
+    chunks: string[];
+  } {
     const serverRoot = getMetroServerRootMemo(projectRoot);
 
     const {
@@ -280,11 +300,12 @@ export function createServerComponentsMiddleware(
         const chunk = context.ssrManifest.get(relativeFilePath);
 
         return {
-          id: relativeFilePath,
+          id: String(createModuleId(file, { platform: context.platform, environment: 'client' })),
           chunks: chunk != null ? [chunk] : [],
         };
       }
 
+      const environment = isServer ? 'react-server' : 'client';
       const searchParams = createBundleUrlSearchParams({
         mainModuleName: '',
         platform: context.platform,
@@ -301,7 +322,7 @@ export function createServerComponentsMiddleware(
         bytecode: false,
         clientBoundaries: [],
         inlineSourceMap: false,
-        environment: isServer ? 'react-server' : 'client',
+        environment,
         modulesOnly: true,
         runModule: false,
       });
@@ -316,6 +337,7 @@ export function createServerComponentsMiddleware(
       clientReferenceUrl.search = searchParams.toString();
 
       const filePath = file.startsWith('file://') ? fileURLToFilePath(file) : file;
+
       const relativeFilePath = path.relative(serverRoot, filePath);
 
       clientReferenceUrl.pathname = relativeFilePath;
@@ -326,9 +348,12 @@ export function createServerComponentsMiddleware(
       }
 
       // Return relative URLs to help Android fetch from wherever it was loaded from since it doesn't support localhost.
-      const id = clientReferenceUrl.pathname + clientReferenceUrl.search;
+      const chunkName = clientReferenceUrl.pathname + clientReferenceUrl.search;
 
-      return { id: relativeFilePath, chunks: [id] };
+      return {
+        id: String(createModuleId(filePath, { platform: context.platform, environment })),
+        chunks: [chunkName],
+      };
     };
   }